fix: separate network gateway logic in a new crate
