+++ Cargo.toml
@@ -2,6 +2,7 @@
 members = [
     ".",
     "crates/smolvm-agent",
+    "crates/smolvm-network",
     "crates/smolvm-napi",
     "crates/smolvm-pack",
     "crates/smolvm-protocol",
@@ -29,6 +30,7 @@ name = "smolvm"
 path = "src/main.rs"
 
 [dependencies]
+smolvm-network = { path = "crates/smolvm-network" }
 smolvm-protocol = { path = "crates/smolvm-protocol" }
 smolvm-pack = { path = "crates/smolvm-pack" }
 smolvm-registry = { path = "crates/smolvm-registry" }
@@ -65,15 +67,6 @@ utoipa-swagger-ui = { version = "9", features = ["axum", "vendored"] }
 ipnet = "2.12.0"
 metrics = "0.24"
 metrics-exporter-prometheus = { version = "0.16", default-features = false }
-crossbeam-queue = "0.3"
-smoltcp = { version = "0.13", default-features = false, features = [
-    "std",
-    "log",
-    "medium-ethernet",
-    "proto-ipv4",
-    "socket-tcp",
-    "socket-udp",
-] }
 
 [dev-dependencies]
 tempfile = "3"
