+++ crates/smolvm-network/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "smolvm-network"
+version = "0.1.0"
+edition = "2021"
+description = "Host-side virtio-net runtime for smolvm"
+license = "Apache-2.0"
+
+[dependencies]
+crossbeam-queue = "0.3"
+humantime = "2"
+libc = "0.2"
+smoltcp = { version = "0.13", default-features = false, features = [
+    "std",
+    "log",
+    "medium-ethernet",
+    "proto-ipv4",
+    "socket-tcp",
+    "socket-udp",
+] }
+tracing = { workspace = true }
