+++ src/network/mod.rs
@@ -1,12 +1,10 @@
 //! Network configuration and backend selection.
 
-pub mod addressing;
 /// Backend selection and serialization helpers.
 pub mod backend;
 /// Launch-time backend planning and request validation rules.
 pub mod launch;
 pub mod policy;
-pub mod virtio;
 
 pub use backend::NetworkBackend;
 pub use launch::{
